requirements.txt
backports.csv==1.0.7
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-crispy-forms==1.8.0
django-fontawesome==1.0
django-import-export==1.2.0
et-xmlfile==1.0.1
get==2019.4.13
gunicorn==19.9.0
idna==2.8
jdcal==1.4.1
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.2.0
post==2019.4.13
public==2019.4.13
pytz==2019.3
PyYAML==5.1.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
South==1.0.2
sqlparse==0.3.0
stripe==2.37.2
tablib==0.13.0
urllib3==1.25.6
whitenoise==4.1.4
xlrd==1.2.0
xlwt==1.3.0
